Add prettier and eslint
