+++ angular/components/.eslintrc.js
@@ -1,0 +1,10 @@
+module.exports =  {
+  parser:  '@typescript-eslint/parser',
+  extends:  [ '@pxblue/eslint-config/ts' ],
+  parserOptions:  {
+    project: "./tsconfig.json",
+  },
+  env: {
+    browser: true
+  }
+};
