+++ angular/components/package.json
@@ -1,14 +1,15 @@
 {
   "name": "components",
   "version": "0.0.0",
-  "main": "./dist",
   "scripts": {
     "ng": "ng",
-    "watch:storybook": "ng build -c storybook --watch",
+    "watch:storybook": "ng build -c storybook",
+    "watch:showcase": "ng build -c showcase",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "prettier": "prettier \"src/**/**.{ts,json,css,scss,html}\" --write",
+    "lint": "eslint \"src/**/**.ts\"",
+    "lint:fix": "eslint \"src/**/**.ts\" --fix"
   },
   "private": true,
   "devDependencies": {
@@ -26,12 +27,18 @@
     "@angular/platform-browser": "~8.2.2",
     "@angular/platform-browser-dynamic": "~8.2.2",
     "@pxblue/colors": "^2.0.0",
+    "@pxblue/eslint-config": "^1.1.3",
     "@pxblue/ng-progress-icons": "^1.1.3",
+    "@pxblue/prettier-config": "^1.0.2",
     "@pxblue/themes": "^3.0.1",
     "@types/jasmine": "~3.5.2",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~13.5.0",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
     "codelyzer": "~5.2.1",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -40,7 +47,7 @@
     "karma-jasmine": "~3.1.0",
     "karma-jasmine-html-reporter": "^1.5.1",
     "ng-packagr": "^5.7.1",
-    "protractor": "~5.4.3",
+    "prettier": "^1.19.1",
     "rxjs": "^6.5.4",
     "ts-node": "~8.6.2",
     "tsickle": "0.38.0",
@@ -48,5 +55,6 @@
     "typescript": "3.5.2",
     "zone.js": "~0.10.2"
   },
-  "dependencies": {}
+  "dependencies": {},
+  "prettier": "@pxblue/prettier-config"
 }
