+++ angular/components/src/channel-value/channel-value.component.scss
@@ -1,0 +1,33 @@
+:host {
+    display: inline-flex;
+}
+.value-wrapper {
+    display: flex;
+}
+.text {
+    font-size: inherit;
+    line-height: 1.2;
+    letter-spacing: 0;
+    margin: 0;
+}
+.units {
+    font-weight: 300;
+}
+.value {
+    font-weight: 600;
+}
+.icon {
+    font-size: inherit;
+    display: block;
+    margin-right: 5px;
+    &:empty {
+        display: none;
+    }
+    ::ng-deep mat-icon {
+        display: block;
+        width: 100%;
+        height: 100%;
+        font-size: 100%;
+        line-height: 1.2;
+    }
+}
