+++ angular/components/src/channel-value/channel-value.module.ts
@@ -3,10 +3,8 @@ import { ChannelValueComponent } from './channel-value.component';
 import { CommonModule } from '@angular/common';
 
 @NgModule({
-  declarations: [ChannelValueComponent],
-  imports: [
-    CommonModule
-  ],
-  exports: [ChannelValueComponent]
+    declarations: [ChannelValueComponent],
+    imports: [CommonModule],
+    exports: [ChannelValueComponent],
 })
-export class ChannelValueModule { }
+export class ChannelValueModule {}
