+++ angular/components/src/empty-state/empty-state.component.html
@@ -1,0 +1,10 @@
+<div class="container" [style.color]="Colors.gray[500]">
+    <div style="line-height: 1">
+        <ng-content select="[empty-icon]"></ng-content>
+    </div>
+    <h2>{{ title }}</h2>
+    <h4 *ngIf="description" [style.color]="Colors.blue[500]">{{ description }}</h4>
+    <div>
+        <ng-content select="[actions]"></ng-content>
+    </div>
+</div>
