+++ angular/components/src/empty-state/empty-state.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { EmptyStateComponent } from './empty-state.component';
+
+@NgModule({
+    declarations: [EmptyStateComponent],
+    imports: [CommonModule],
+    exports: [EmptyStateComponent],
+})
+export class EmptyStateModule {}
