+++ angular/components/src/hero-banner/hero-banner.component.ts
@@ -1,0 +1,16 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+    selector: 'pxb-hero-banner',
+    template: `
+        <div class="banner">
+            <ng-content select="pxb-hero"></ng-content>
+        </div>
+        <mat-divider class="divider" *ngIf="divider"></mat-divider>
+    `,
+    styleUrls: ['./hero-banner.component.scss'],
+    inputs: ['divider'],
+})
+export class HeroBannerComponent {
+    divider: boolean = true;
+}
