+++ angular/components/src/hero-banner/public-api.ts
