+++ angular/components/src/hero/hero.component.scss
@@ -1,0 +1,58 @@
+@import '../../../node_modules/@pxblue/colors/palette';
+:host {
+    flex: 1 1 0px;
+}
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    overflow: hidden;
+    color: map-get($pxb-gray, 500);
+    padding: 16px 8px;
+}
+
+// Icon Styles
+.icon {
+    margin-bottom: 8px;
+    text-align: center;
+    color: map-get($pxb-gray, 800);
+}
+:host ::ng-deep .icon mat-icon {
+    display: block;
+    width: 100%;
+    height: 100%;
+    font-size: 100%;
+    line-height: 1.2;
+    text-align: center;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    svg {
+        transform: scale(1.5);
+        transform-origin: top left;
+    }
+}
+:host ::ng-deep .icon.large mat-icon svg {
+    transform: scale(3);
+}
+
+// Channel Value Styles
+.channel-value {
+    display: flex;
+    color: map-get($pxb-gray, 800);
+}
+
+// Label Styles
+.label {
+    font-size: inherit;
+    line-height: 1.2;
+    letter-spacing: 0;
+    font-weight: 600;
+    width: 100%;
+    text-align: center;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    margin: 0;
+}
