+++ angular/components/src/index.ts
@@ -1,5 +1,4 @@
-export * from './lib/channel-value/public-api';
-export * from './lib/empty-state/public-api';
-export * from './lib/hero/public-api';
-export * from './lib/hero-banner/public-api';
-
+export * from './channel-value/public-api';
+export * from './empty-state/public-api';
+export * from './hero/public-api';
+export * from './hero-banner/public-api';
